requirements.txt
Flask
python-docx
pdf2docx
